requirements.txt
hydra-core
tqdm
wandb
pre-commit
black
gpustat
loguru
einops==0.7.0
fancy-einsum==0.0.3
huggingface-hub==0.21.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
omegaconf==2.3.0
safetensors==0.4.2
tokenizers==0.15.2
huggingface_hub
numpy
termcolor
evaluate
mauve-text
git+https://git@github.com/jdeschena/simple-cfg
git+https://github.com/EleutherAI/lm-evaluation-harness
sentencepiece
seaborn
openai
blobfile
SciencePlots
datasets==2.18.0
fsspec==2024.2.0
git-lfs==1.6
lightning==2.2.1
notebook==7.1.1
nvitop==1.3.2
packaging==23.2
pandas==2.2.1
rich==13.7.1
scikit-learn==1.4.0
timm==0.9.16
transformers==4.38.2
wandb==0.13.5
jupyter
accelerate
pyyaml
ninja
aim
vendi_score[all]
